Added a new "ratio" method that relies on the logarithmic method
Technically, just like the sum fucntion, the log_product() function was
renamed to log() and now takes a char parameter that specifies '*' for
multiplication and '/' for the sum. In calcualting the delta, i am
unsure if the x average should be multiplied by the sum of the 2
relative uncerts or their difference. As a conservative measure, i've
kept the sum for now (an inaccurately larger delta is better than an
inaccurately smaller delta).

Signed-off-by: Omar Mahboubi <omar_mahboubi@tutanota.com>