methods.cpp: Confirmed uncerainty of the ratio
in commit c35e5ac0, a new "ratio" method was introduced, but at the time
i was unsure if the relative uncertainty of z was the sum of relative
uncertainties of a and b, or their difference. I confirmed it was the
former: uncertainty can't be subtracted. No real change to the logic was
intorduced; just removed the TODO comment.

Signed-off-by: Omar Mahboubi <omar_mahboubi@tutanota.com>